Mostly updates to the automated download and running code.
MODIFIED FILES:
    Downloading/get_omno2.sh - added a safety catch that will exit and email the
admin if the download notice isn't found after six hours. This should help
prevent hangs if the download notice is on a different day.

    Downloading/run_read_omno2.sh - had to adjust the way of finding the last
OMNO2 file to allow it to be insensitive to the version number.

NEW FILES:
    Downloading/automessage.sh - this is the bash script that allows the
automatic emailing of the admin. Uses the sendEmail package, installable through
APT on Ubuntu.

    Downloading/run_behr_main.sh - shell script like run_read_omno2.sh that can
be used to run BEHR_main.m (unsurprisingly).

    Downloading/run_publishing.sh - runs the BEHR_publishing_v2.m to produce the
HDF and TXT files for the website.  Also ensures that the website's BEHR version
text file is up-to-date.